Backport #11196 to 7.0: Import updated ECS 1.0.0 go library. (#11196) (#11206)
* Import updated ECS 1.0.0 go library.
* Adjust Packetbeat to use int64 for URL ports.
* Re-generate golden files for filebeat and x-pack/filebeat

+++ packetbeat/protos/http/event.go
@@ -73,7 +73,7 @@ type ProtocolFields struct {
 }
 
 // netURL returns a new ecs.Url object with data from the HTTP request.
-func newURL(host string, port int32, path, query string) *ecs.Url {
+func newURL(host string, port int64, path, query string) *ecs.Url {
 	u := &ecs.Url{
 		Scheme: "http",
 		Domain: host,
@@ -87,7 +87,7 @@ func newURL(host string, port int32, path, query string) *ecs.Url {
 	return u
 }
 
-func synthesizeFullURL(u *ecs.Url, port int32) string {
+func synthesizeFullURL(u *ecs.Url, port int64) string {
 	if u.Domain == "" || port <= 0 {
 		return ""
 	}
